+++ sdk/src/types/block/output/unlock_condition/storage_deposit_return.rs
@@ -69,10 +69,10 @@ pub mod dto {
     use crate::types::block::{address::dto::AddressDto, error::dto::DtoError};
 
     #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+    #[serde(rename_all = "camelCase")]
     pub struct StorageDepositReturnUnlockConditionDto {
         #[serde(rename = "type")]
         pub kind: u8,
-        #[serde(rename = "returnAddress")]
         pub return_address: AddressDto,
         pub amount: String,
     }