+++ sdk/tests/wallet/message_interface.rs
@@ -173,7 +173,7 @@ async fn message_interface_events() -> Result<()> {
                 account_id: "alias".into(),
                 method: AccountMethod::RequestFundsFromFaucet {
                     url: FAUCET_URL.to_string(),
-                    address: account.public_addresses[0].address().to_bech32(),
+                    address: account.public_addresses[0].address().to_string(),
                 },
             };