+++ sdk/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.3.0 - 2023-XX-XX
+
+### Changed
+
+- `AddressWrapper` renamed to `Bech32Address` and moved to `types`;
+- `Address::try_from_bech32_with_hrp` address and HRP return have been reversed;
+- `PostBlockPayload::payload_dto` renamed to `payload`;
+- `SendNativeTokens::addresses_native_tokens` renamed to `addresses_and_native_tokens`;
+- `SendNft::addresses_nft_ids` renamed to `addresses_and_nft_ids`;
+
 ## 0.2.0 - 2023-04-17
 
 ### Added